requirements.txt
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10
openai==1.12.0
python-dotenv==1.0.0
fastapi==0.110.0
uvicorn==0.29.0
faiss-cpu==1.7.4
pypdf2==3.0.1
pillow==10.2.0
pytesseract==0.3.10
